Add icon for back button
We got the back button for free from the react stack navigation library,
but this was customizing the back button to have the icon. This was an
initial POC, as it currently only works on Screen 1 and would need to be
copy-pasted to count for other pages.

Next step is to add this to a class that our screens are derived from.