Add simple placeholder for tests, install dependencies for Jest
This is simply the addition of a single unit test for App.js --- it's
really simplistic and doesn't actually test much functionality, but we
want to get it uploaded so we can get started on the GitLab work.

Included are the changes to package.json and package-lock.json to enable
running tests and install the dependencies for Jest.